requirements.txt
torch~=2.4.1
Flask~=3.0.3
transformers~=4.44.2
torchvision~=0.19.1
pillow~=10.4.0
sentencepiece~=0.2.0
protobuf~=5.28.1